requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.12.0
pillow==10.2.0
python-dotenv==1.0.1
lxml==5.1.0
validators==0.22.0
